==> 1. pre-processing.ipynb <==
# Import libs

import pandas as pd

# Read the data

users_col_names = ['ID', 'Type', 'Name', 'Reserved field', 'Description']
users = pd.read_csv('raw_data/users.csv', sep='|', on_bad_lines='skip', names=users_col_names)

subscriptions_col_names = ['FollowerID', 'FollowedID']
subs = pd.read_csv('raw_data/subscriptions.csv', sep=',', on_bad_lines='skip', names=subscriptions_col_names)

services_col_names = ['UserID', 'ServiceID', 'ServiceName', 'ServiceURL', 'UserNameOnService', 'UserURLOnService']
services = pd.read_csv('raw_data/services.csv', sep='|', on_bad_lines='skip', names=services_col_names)

likes_col_names = ['userID', 'PostID', 'Timestamp']
likes = pd.read_csv('raw_data/likes.csv', sep='\t', on_bad_lines='skip', names=likes_col_names)

following_col_names = ['FollowerID', 'FollowedID', 'Timestamp']
following = pd.read_csv('raw_data/followingAugSept.csv', sep='\t', on_bad_lines='skip', names=following_col_names)

comments_col_names = ['PostID', 'EntryID', 'PostedBy', 'SourceName', 'SourceURL', 'GeoX', 'GeoY', 'Timestamp', 'Text', 'NumImg', 'ImgURL', 'NumVid', 'VidURL']
comments = pd.read_csv('raw_data/commentAugSept.csv', sep='\t', on_bad_lines='skip', names=comments_col_names)

entries_col_names = ['PostID', 'PostedBy', 'SourceName', 'SourceURL', 'GeoX', 'GeoY', 'Timestamp', 'Text', 'NumImg', 'ImgURL', 'NumVid', 'VidURL']
entries1 = pd.read_csv('raw_data/entries1.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
entries2 = pd.read_csv('raw_data/entries2.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
entries3 = pd.read_csv('raw_data/entries3.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
entries = pd.concat([entries1, entries2, entries3], ignore_index=True)

# Drop unrelevant columns

users.drop('Reserved field', axis=1, inplace=True)

services.drop(['ServiceName', 'UserNameOnService', 'UserURLOnService'], axis=1, inplace=True)

comments.drop(['SourceName', 'GeoX', 'GeoY', 'NumImg', 'NumVid', 'ImgURL', 'VidURL'], axis=1, inplace=True)

entries.drop(['SourceName', 'GeoX', 'GeoY', 'NumImg', 'NumVid', 'ImgURL', 'VidURL'], axis=1, inplace=True)

# Get data overview

## Users

users.head()

users.info()

users.describe()

## Subs

subs.head()

subs.info()

subs.describe()

## Services

services.head()

services.info()

services.describe()

## Likes

likes.head()

likes.info()

likes.describe()

## Following 

following.head()

following.info()

following.describe()

## Comment

comments.head()

comments.info()

comments.describe()

## Entries 1,2,3

entries1.head()

entries1.info()

entries1.describe()

entries2.head()

entries2.info()

entries2.describe()

entries3.head()

entries3.info()

entries3.describe()

# Data exploration

## Users

users[users['Name'].isna()]

users.drop_duplicates(subset=['ID', 'Type'], inplace=True)

users.head()

## Subs

subs[subs.duplicated()].head()

subs[(subs['FollowerID']=='cheteam') & (subs['FollowedID']=='104455115')]

subs.drop_duplicates(subset=['FollowerID', 'FollowedID'], inplace=True)

subs.shape

## Services

services.dropna(subset=['UserID'], inplace=True)

services.head()

services.groupby(['ServiceID', 'ServiceURL']).count()

services[services.duplicated()].head()

services.drop_duplicates(subset=['UserID', 'ServiceID'], inplace=True)

## Likes

likes.head()

likes['Timestamp'] = pd.to_datetime(likes['Timestamp']).dt.tz_localize('Etc/GMT-1').dt.tz_convert('UTC').dt.strftime('%Y-%m-%dT%H:%M:%S')

likes[likes.duplicated()].head()

likes.drop_duplicates(subset=['userID', 'PostID', 'Timestamp'], inplace=True)

## Following

following.head()

def convert_timestamp(row):
    timestamp = row['Timestamp']
    timestamp = timestamp.split('.')
    timestamp = timestamp[0]
    return timestamp

following['Timestamp'] = following.apply(convert_timestamp, axis=1)
following['Timestamp'] = pd.to_datetime(following['Timestamp']).dt.tz_localize('Etc/GMT-1').dt.tz_convert('UTC').dt.strftime('%Y-%m-%dT%H:%M:%S')

following.head()

## Comment

comments.head()

comments[['EntryID', 'PostedBy', 'Timestamp']].duplicated().sum()

comments[comments[['EntryID', 'PostedBy', 'Timestamp']].duplicated()].head()

comments.drop_duplicates(subset=['EntryID', 'PostedBy', 'Timestamp'], inplace=True)

comments['Timestamp'] = pd.to_datetime(comments['Timestamp']).dt.tz_localize('Etc/GMT-1').dt.tz_convert('UTC').dt.strftime('%Y-%m-%dT%H:%M:%S')

## Entries 1, 2, 3

entries = pd.concat([entries1, entries2, entries3])
entries.shape

entries.head()

entries['Timestamp'] = pd.to_datetime(entries['Timestamp']).dt.tz_localize('Etc/GMT-1').dt.tz_convert('UTC').dt.strftime('%Y-%m-%dT%H:%M:%S')

entries.dropna(subset=['Timestamp'], inplace=True)

entries.drop_duplicates(subset=['Text', 'SourceURL', 'PostedBy', 'Timestamp'], inplace=True)

## Remove timestamp outliers

likes = likes[likes['Timestamp'] >= '2010-01-01']
following = following[following['Timestamp'] >= '2010-01-01']
comments = comments[comments['Timestamp'] >= '2010-01-01']
entries = entries[entries['Timestamp'] >= '2010-01-01']

# Save pre-processed files

users.to_parquet('processed_data/users.parquet', index=False)
subs.to_parquet('processed_data/subs.parquet', index=False)
services.to_parquet('processed_data/services.parquet', index=False)
likes.to_parquet('processed_data/likes.parquet', index=False)
following.to_parquet('processed_data/following.parquet', index=False)
comments.to_parquet('processed_data/comments.parquet', index=False)
entries.to_parquet('processed_data/entries.parquet', index=False)
