==> 2. visualization.ipynb <==
import pandas as pd
from ydata_profiling import ProfileReport

users_col_names = ['ID', 'Type', 'Name', 'Reserved field', 'Description']
raw_users = pd.read_csv('raw_data/users.csv', sep='|', on_bad_lines='skip', names=users_col_names)
users = pd.read_parquet('processed_data/users.parquet', engine='pyarrow')

raw_users_profile = ProfileReport(raw_users, title="Raw Users Report", explorative=True)
raw_users_profile.to_file("profiling/raw_users_profile.html")

users_profile = ProfileReport(users, title="Processed Users Report", explorative=True)
users_profile.to_file("profiling/processed_users_profile.html")

subscriptions_col_names = ['FollowerID', 'FollowedID']
raw_subs = pd.read_csv('raw_data/subscriptions.csv', sep=',', on_bad_lines='skip', names=subscriptions_col_names)
subs = pd.read_parquet('processed_data/subs.parquet', engine='pyarrow')

raw_subs_profile = ProfileReport(raw_subs, title="Raw Subs Report", explorative=True)
raw_subs_profile.to_file("profiling/raw_subs_profile.html")

subs_profile = ProfileReport(subs, title="Processed Subs Report", explorative=True)
subs_profile.to_file("profiling/processed_subs_profile.html")

services_col_names = ['UserID', 'ServiceID', 'ServiceName', 'ServiceURL', 'UserNameOnService', 'UserURLOnService']
raw_services = pd.read_csv('raw_data/services.csv', sep='|', on_bad_lines='skip', names=services_col_names)
services = pd.read_parquet('processed_data/services.parquet', engine='pyarrow')

raw_services_profile = ProfileReport(raw_services, title="Raw Services Report", explorative=True)
raw_services_profile.to_file("profiling/raw_services_profile.html")

services_profile = ProfileReport(services, title="Processed Services Report", explorative=True)
services_profile.to_file("profiling/processed_services_profile.html")

likes_col_names = ['userID', 'PostID', 'Timestamp']
raw_likes = pd.read_csv('raw_data/likes.csv', sep='\t', on_bad_lines='skip', names=likes_col_names)
likes = pd.read_parquet('processed_data/likes.parquet', engine='pyarrow')

raw_likes_profile = ProfileReport(raw_likes, title="Raw Likes Report", explorative=True)
raw_likes_profile.to_file("profiling/raw_likes_profile.html")

likes_profile = ProfileReport(likes, title="Processed Likes Report", explorative=True)
likes_profile.to_file("profiling/processed_likes_profile.html")

following_col_names = ['FollowerID', 'FollowedID', 'Timestamp']
raw_following = pd.read_csv('raw_data/followingAugSept.csv', sep='\t', on_bad_lines='skip', names=following_col_names)
following = pd.read_parquet('processed_data/following.parquet', engine='pyarrow')

raw_following_profile = ProfileReport(raw_following, title="Raw Following Report", explorative=True)
raw_following_profile.to_file("profiling/raw_following_profile.html")

following_profile = ProfileReport(following, title="Processed Following Report", explorative=True)
following_profile.to_file("profiling/processed_following_profile.html")

comments_col_names = ['PostID', 'EntryID', 'PostedBy', 'SourceName', 'SourceURL', 'GeoX', 'GeoY', 'Timestamp', 'Text', 'NumImg', 'ImgURL', 'NumVid', 'VidURL']
raw_comments = pd.read_csv('raw_data/commentAugSept.csv', sep='\t', on_bad_lines='skip', names=comments_col_names)

raw_comments_profile = ProfileReport(raw_comments, title="Raw Comments Report", explorative=True)
raw_comments_profile.to_file("profiling/raw_comments_profile.html")

comments = pd.read_parquet('processed_data/comments.parquet', engine='pyarrow')

comments_profile = ProfileReport(comments, title="Processed Comments Report", explorative=True)
comments_profile.to_file("profiling/processed_comments_profile.html")

entries_col_names = ['PostID', 'PostedBy', 'SourceName', 'SourceURL', 'GeoX', 'GeoY', 'Timestamp', 'Text', 'NumImg', 'ImgURL', 'NumVid', 'VidURL']
entries1 = pd.read_csv('raw_data/entries1.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
entries2 = pd.read_csv('raw_data/entries2.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
entries3 = pd.read_csv('raw_data/entries3.csv', sep='\t', on_bad_lines='skip', names=entries_col_names)
raw_entries = pd.concat([entries1, entries2, entries3], ignore_index=True)
entries = pd.read_parquet('processed_data/entries.parquet', engine='pyarrow')

raw_entries_profile = ProfileReport(raw_entries, title="Raw Entries Report", explorative=True)
raw_entries_profile.to_file("profiling/raw_entries_profile.html")

entries_profile = ProfileReport(entries, title="Processed Entries Report", explorative=True)
entries_profile.to_file("profiling/processed_entries_profile.html")
